[BUGFIX] Dispatch native `change` event in TBE_EDITOR.fieldChanged()
With #87324 some code of FormEngine was moved into separated modules. In
the very same patch, most of the code was rewritten to use native
JavaScript as much as possible. However, with this change,
`TBE_EDITOR.fieldChanged()` was partially broken as it uses jQuery's
`triggerHandler` method to trigger a specific event. Events that are
registered via jQuery's `on()` method work fine, but the approach is
incompatible with JavaScript's native `addEventListener()` introduced
in the mentioned change.

As a quick fix, `TBE_EDITOR.fieldChanged()` now triggers the same event
for native JavaScript code via `dispatchEvent()`.

Resolves: #88738
Related: #87324
Releases: master
Change-Id: I38fbb97f86f6765a45ad763c27e7afdac5754b1c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61278
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>