[!!!][TASK] Remove cHash calculation in indexed search
cHash is now added automatically when a URL is generated.
The relevant query parameters when indexing
are stored in "static_page_arguments", which allows
to remove the database field "cHashParams".

Therefor it is not necessary anymore to configure if
cHash should be taken into account when creating
a configuration for indexed search.

Instead, when linking to a page on a search result,
the page arguments are added. In addition, when
pages are indexed, only the static page arguments
are evaluated.

Debug Information when indexing is also adding
data more sensibly via json_encode/decode.

Effectively, this means that specific handling
for cHash resolving is fully removed from
EXT:indexed_search.

Related: #87193
Resolves: #88741
Releases: master
Change-Id: I84738612d42615a3ac24d271c5509b52467d81af
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61284
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>