[!!!][TASK] Always concatenate all given files in ResourceCompressor
ResourceCompressor allows to only concatenate files
given in a given "baseDirectory" as separate option, which
was used to only include files that are registered via TBE_STYLES
for backend purposes.

The functionality is removed to always concatenate all CSS files
in TYPO3 Backend.

A heavy dependency to DocumentTemplate of ALL backend
functionality using PageRenderer is therefore removed.

Resolves: #88758
Releases: master
Change-Id: I07e2f15a3dd2298371db87ce2723ded0f6a56f31
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61296
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>