[BUGFIX] Explicitly cast variable to string
Class \TYPO3\CMS\Extbase\Core\Bootstrap calls method
\TYPO3\CMS\Extbase\Mvc\ResponseInterface::shutdown()
which does not declare a strict return type yet.

Thus, it's possible that the returned value is not
of type string which then violates the return type
promise of method handleRequest() of the Bootstrap
class.

Therefore, the return value is explicitly casted
to string.

Releases: master
Resolves: #88829
Change-Id: Ibea2451e87c5c66070e5ce4ed5f30c408b7412cd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61345
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Armin Vieweg <armin@v.ieweg.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>