[BUGFIX] Streamline routing redirects to default site language
Calling the frontend with an URL that does not contain a valid base
URI for a configured language resulted in a temporary redirect (307)
to the base URI of the default language. In order to allow detecting
outdated links returning a page not found (404) is used.

Example: https://example.org/en/ is the base URI of a valid language
+ https://example.org/ -> redirects to default language /en/ (307)
+ https://example.org/nothing/ -> responds a page not found  (404)

Releases: master, 9.5
Resolves: #88838
Change-Id: I9a3eeb53da8e0bb92799d8e29404513699411078
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61346
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Chris Müller <typo3@krue.ml>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Chris Müller <typo3@krue.ml>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>