[BUGFIX] Ensure ReflectionService can be used during TCA load
In order to support the 'tcaIsBeingBuilt' for legacy signal
listeners that are *not* available in the symfony container,
we need to enable the Extbase Object Container to use the
(lazy loaded) ReflectionService, which requires the CacheManager
to be available.
Therefore we now mark "boot.state" done once ext_localconf.php
has been loaded. boot.state = true will allow the CacheManager
to be created in TYPO3\CMS\Core\ServiceProvider instead of
throwing a LogicException.

Releases: master
Resolves: #88843
Change-Id: I93f182d1e84f65ae005a8e51a0faff5e47dc36f0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61373
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>