[BUGFIX] Make sure AdditionalConfiguration is loaded
This loads AdditionalConfiguration.php even if LocalConfiguration.php
could not be loaded.

Before this change, configuration for the install tool could only be
given if a LocalConfiguration was present.

With this change, AdditionalConfiguration (if present) is applied
on top of DefaultConfiguration.

In a FIRST_INSTALL scenario, this enables (among others) the correct
evaluation of a set trustedHostsPattern in AdditionalConfiguration.

Resolves: #88846
Resolves: #86356
Releases: master, 9.5
Change-Id: I6c0e39eb57ce6b28b493f1ae9716b478b643dda9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61363
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Jonas Eberle <flightvision@googlemail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>