[TASK] Deprecate BackendUtility::editOnClick()
This patch deprecates BackendUtility::editOnClick() which is used to build
a JavaScript-based redirect to the EditDocumentController for use in an
`onclick` attribute. This approach is not necessary anymore, thus all
links use a real `href` attriute now.

Also, this patch removes the following methods:
- TYPO3\CMS\Backend\View\PageLayoutView::linkWrapTable()
- TYPO3\CMS\Backend\View\PageLayoutView::linkWrapItems()
- TYPO3\CMS\Backend\View\PageLayoutView::newContentElementOnClick()

Reason is that these methods contain code that would need a migration, but
are unused in the whole Core. Since the class is marked as "internal",
it's considered safe to remove them.

Resolves: #88787
Releases: master
Change-Id: I8bd3868287204fc3275a43f938d9aeb9fbb6306d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61377
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Windloff
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Daniel Windloff
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>