[BUGFIX] Make redis pconnect calls unique
By using the persistent_id parameter of the redis->pconnect
method, we ensure that the connection is not shared between
multiple Redis*Backends connecting to the same Redis server.

Omitting the persistent_id causes the same connection to be reused
whenever another Redis*Backend is created, whereby the last
connection selects the database to use for the connection, effectively
causing all Redis*Backends to write to the same database.

The RedisSessionBackend uses the pconnect method by default and
therefore requires this fix in order to distinguish FE and BE
backends correctly, if both are stored within a Redis database.

The pconnect is optional for the cache RedisBackend, but we still
use the database number now for the persistent_id parameter.

Resolves: #88866
Releases: master, 9.5, 8.7
Change-Id: I987c36e89f2ab53fd5177cdc7051811b116bcad0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61393
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>