[FEATURE] Allow array of handlers in RequestFactory
Guzzle has the possibility to push custom middleware handlers during
client initialization. With this patch it is now possible to define own
handlers as an array of callables in $GLOBALS['TYPO3_CONF_VARS']['HTTP']['handler'].
The RequestFactory builds a HandlerStack from that array
and injects it into the created client.

Resolves: #88871
Releases: master
Change-Id: I62cb9c58c9ef7f8ce66255eb34b3a0953666cbc2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61407
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>