[BUGFIX] Allow sys_file_metadata to be translatable
Since sys_file_metadata records are stored on pid=0,
there is no site configuration available. This situation is
similar to be_users.limitToLanguage feature, where
all languages of all sites need to be visible (depending
on the backend user rights).

This patch adds an option for the TranslationConfigurationProvider
to render all languages of all sites as a combined list
if the given pageId is 0 (based on the current user rights).

Resolves: #88897
Releases: master
Change-Id: I14b1bb6ae5c8577571cf5fa7bc2da1c1824c0f58
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61419
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>