[BUGFIX] Set module menu state on render time
The module menu now contains its state whether either the menu itself, or
its module groups are expanded or collapsed. This information is now taken
into account when the markup of the backend viewport is rendered.

With this patch, some AJAX requests done after the backend is initialized
can bre dropped, improving the backend UX since the state of the menus is
correctly handled on render time.

Resolves: #88935
Releases: master, 9.5
Change-Id: I3f3765926d9d3f406af0ee0bc29993d547771853
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61477
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Richard Haeser <richard@maxserv.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Reviewed-by: Richard Haeser <richard@maxserv.com>