[BUGFIX] Disable copying taboverride due to changed source map
With #86804 the hardcoded sourcemap reference got removed from
taboverrides. Unfortunately, executing `grunt build` copies the original
file from node_modules and basically reverts that very change. For this
reason the file is excluded from the "copy" task to keep the current
state of the file as is.

Resolves: #88999
Related: #86804
Releases: master, 9.5, 8.7
Change-Id: I53e9d39b979ed239bcac96a8d05b7b843622657b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61540
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>