[BUGFIX] Allow selective saving in SiteConfiguration
YAML SiteConfiguration allows imports and custom configuration
that is not configurable via UI to be written directly into the
configuration files, however, when saving with the UI these
settings are overwritten and resolved. This change changes the
SiteConfiguration in a way, that only modified values are written
and imports are completely left as they are.

Resolves: #86803
Related: #89116
Releases: master, 9.5
Change-Id: I92ece4959d57f41535e6fc8c3c7c9d2894d7f920
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61603
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>