[BUGFIX] Handle base paths with mixed FQDN and "/"
If a site has "/" as its entrypoint but the languages
have FQDNs or subdomains or a mixture between subdomain
the base without domain always gets found and the subdomains
are not considered.

Currently this situation is handled in
  SiteMatcher->getRouteCollectionForAllSites()
which does not fit a certain scheme if there is a mixture.

This change corrects this and adds tests to handle any cases:

- just / as base, and then /de and /en as language base
- just / as base, and then subdmoains or full domains+path as language base
- a full URL as base, and then path parts as language base
- a full URL as base and then full URLs + subdomains as language base
- Also check for http and http variants.

Resolves: #89183
Releases: master, 9.5
Change-Id: If570118ac371e1a1589f064605c0176a53220f0f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61650
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Daniel Gorges <daniel.gorges@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Daniel Gorges <daniel.gorges@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>