[FEATURE] Create sites on page creation
A new hook is introduced that creates a site configuration when a new
page on root level is created. This takes effect for pages of type
"default", "link" and "shortcut".

To reduce the likelihood for conflicts, a shortened MD5 hash of the page
id appended to the site identifier.

Resolves: #89142
Releases: master
Change-Id: Ibe2957e3789f2a165e36949ae5fb4fa2a1a572df
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61658
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>