[BUGFIX] Keep workspace ID when discarding versioned records
When a record was modified in a workspace, and then discarded, TYPO3 previously
set the t3ver_wsid to "0", which basically meant "we release it to live workspace
as a offline version with pid=-1".

However, this turns out to be ugly, as this information is then floating in live workspace,
without any information where this record was from.

Instead, "discarding versioned records" now keeps the t3ver_wsid=X, and just marks
the versioned record as "deleted" - or removes it completely if the database table
does not support to mark records as deleted.

As a result, there will be no records in live workspace anymore with "pid=-1" in the future
anymore. To remove any "old" discarded records, an upgrade wizard will follow
in a followup patch.

Resolves: #89166
Releases: master
Change-Id: I8ccab3cd2053c27d9b0ecd9f171a83b9097f29dd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61671
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Daniel Gorges <daniel.gorges@b13.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Daniel Gorges <daniel.gorges@b13.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>