Examples from Manual should be working
Fields without softref are not being checked. Since pages:media does not have Softref set in its TCA it will never get checked.  Field pages:url has an unresolved issue since 6 years. So better don't mention it as an example for configuration until it works. All of this has been the same since at least TYPO3 8.7. So the information can be backported.
The default configuration also contains the two not working fields. It causes users and administrators to feel wrongly in safety. We had a non-working External Link in the menu of our customer (level 4 so not easy to detect) and trusted on being notified if something in pages:url is not working.