[BUGFIX] Generate notification action on rendering
Due to scoping issues if a notification is generated in an iframe, the
actions are now generated when the notification is rendered.

This requires a change how the payload is passed to the notification,
instead of the action instance it now contains the action type and the
callback. The action instance is generated by a factory class.

To bypass issues with browser's garbage collection that kicks in when an
iframe changes, the callback methods are stringified and immediately
recovered by eval().

Resolves: #89173
Releases: master
Change-Id: I45624a26bad1527f9d6222ad10e1f9a1384ee07e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61706
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Tobi Kretschmann <tobi@tobishome.de>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Tobi Kretschmann <tobi@tobishome.de>
Reviewed-by: Steffen Frese <steffenf14@gmail.com>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>