[BUGFIX] Re-apply sorting after workspaces in FileRepository
When in workspace mode, overlays are made but if there is
a different sorting (reordering of items in a workspace), the
FileRepository does not consider this, as the workspace
overlays are applied later-on.

As this patch is making Workspaces with FAL finally usable, and is
non-intrusive, and fixes a minor data / repository
issue, I consider backporting this down to 8.7 as well.

Resolves: #71973
Releases: master, 9.5, 8.7
Change-Id: Ie230558242e3ebfef541e8eb969fa21dd15708af
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61733
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>