[BUGFIX] Allow slug recreation if postModifiers are defined
Slug fields currently only support slug recreation if the given
generator fields are directly visible in the backend. However, this
makes it impossible to recreate slugs even if some postModifiers are
defined who can take care of successful slug generation.

This commit allows slug recreation if at least one postModifier is
defined, even if the defined generator fields are not usable for
slug generation.

Therefore also EditDocumentController was adjusted to not
automatically add slug fields to columnsOnly in case postModifiers
are defined for this record.

Resolves: #89187
Releases: master, 10.4
Change-Id: Ifd50ec106b5b302cea23d37e470eafae840c8993
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61716
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>