[BUGFIX] Ensure correct site is used when redirecting
The RedirectService recives a NullSite instance from the current
request. This patch checks whether the redirect target is a page and
fetches its site configuration to be able to create a proper TSFE
instance.

Also, the TSFE instance now does not fetch any template configuration
anymore as it is rather extraneous for a redirect.

Resolves: #88877
Releases: master
Change-Id: If0463d3ad9e8af47950f001c8646dc8a15b6a02b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61838
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>