[BUGFIX] Page preview respects language fallbacks
When a default language page is hidden, it should also be visible
when a translated page is rendered but not available.

The current "simple" check to only check for a specific language
to find out if a page is hidden, is wrong, as l18n_cfg could be built
in a way that a page translation (language>0) does not exist, but
the default record is actually available.

This change overcomes a long-standing issue that
translations of hidden pages can be previewed.

Resolves: #89130
Releases: master, 9.5
Change-Id: I8bc8f1e55a07b680c7b7b9591ec2936d50840548
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61842
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Tobias Gaertner <tobias.gaertner@benaja-websolutions.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Tobias Gaertner <tobias.gaertner@benaja-websolutions.com>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>