[TASK] Add configuration for Crowdin
Crowdin requires a basic configuration which defines where the
XLIFF files and their translations are located. The configuration is used
by the GitHub integration which is offered by Crowdin itself.
See https://support.crowdin.com/github-integration/ for more information
about that.

Even though the translations will not be pushed back to the GitHub
repository of TYPO3 (as pull requests), the "translation" property is
required as the manual export of translations will use it.

Resolves: #89281
Releases: master, 9.5
Change-Id: I1a5f94b416a0652081976bcdbd9ea7d61fb202dc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61826
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: TYPO3com <noreply@typo3.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>