[BUGFIX] Reset belog search constraints on memory exhaustion
If a backend user uses the search in belog with no boundaries or sets a
custom date limit with boundaries that are too wide, the script
execution will experience a "Allowed memory of ... exhausted ..."
fatal error in PHP.

Because the search constraints are saved in the backend users uC
nonetheless, any further effort to access the belog module would result
in a fatal error, thus making it unusable.

This patch makes the belog module detect, if the script execution fails
because of an exhausted memory and reset the search constraints to its
initial values.

Resolves: #89256
Releases: master, 9.5
Change-Id: I08f60399aed3a209cf98f5b33e311d22288e05ce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61878
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: Steffen Frese <steffenf14@gmail.com>
Tested-by: Sascha Rademacher <sascha.rademacher+typo3@gmail.com>
Tested-by: Tobi Kretschmann <tobi@tobishome.de>
Tested-by: Julian Geils <j_geils@web.de>
Tested-by: Felix <f.pachowsky@neusta.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Steffen Frese <steffenf14@gmail.com>
Reviewed-by: Sascha Rademacher <sascha.rademacher+typo3@gmail.com>
Reviewed-by: Tobi Kretschmann <tobi@tobishome.de>
Reviewed-by: Julian Geils <j_geils@web.de>
Reviewed-by: Felix <f.pachowsky@neusta.de>