[BUGFIX] Fix 'pidonly' check in backend search
The TCA setting 'pidonly', which is used to limit a field to be only
searched in the single page view in the list module, was broken and
didn't behave as described in the official TCA documentation.

First, the 'pidonly' setting is expected to be a boolean. However, the
'if' condition would fail for a boolean because the in_array() check in
the code was using strict type mode. Also, this check would never care
about the actual value of 'pidonly', only if it was set.

Second, you couldn't search for an integer value, while not using the
'pidonly' setting, at all.

This patch fixes this behaviour.

Resolves: #89002
Releases: master, 9.5
Change-Id: Iedac044c32b8dd500f4de175ae97ab17984725f8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61884
Reviewed-by: Julian Geils <j_geils@web.de>
Reviewed-by: Sascha Rademacher <sascha.rademacher+typo3@gmail.com>
Reviewed-by: Felix P. <f.pachowsky@neusta.de>
Reviewed-by: Henning Liebe <h.liebe@neusta.de>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Sascha Rademacher <sascha.rademacher+typo3@gmail.com>
Tested-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>