[BUGFIX] Fix "Show" link in the context menu of page translations
In the list module, when opening the context menu of a
"Page Translation" record, a "Show" link will be shown, which is
supposed to open the page in the frontend with the correct language.
However this link was broken, because the PageProvider class didn't
respect page translations at all.

Resolves: #56351
Releases: master, 9.5
Change-Id: I46800ac912d5a964b236c615099a0537227baf82
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61887
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Riccardo De Contardi <erredeco@gmail.com>
Tested-by: Sascha Rademacher <sascha.rademacher+typo3@gmail.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Sascha Rademacher <sascha.rademacher+typo3@gmail.com>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>