[BUGFIX] Prevent broken AND where statement in cObj->searchWhere()
If searchWhere() in ContentObjectRenderer is called with search words
that are smaller than 3 characters in length, a broken " AND " where
statement will be returned, that, if put inside a database query, would
lead to a fatal error. In such a case, return an empty string to prevent addWhere call.

Resolves: #87620
Releases: master, 9.5
Change-Id: Ib75b2fe8f20fb5b3c956976ff7aae014d0eb784e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61930
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: Christian Eßl <indy.essl@gmail.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Reviewed-by: Christian Eßl <indy.essl@gmail.com>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>