[FEATURE] Use SiteLanguageAwareInterface to check for trait existence
In AspectFactory::enrich(), the function would use class_uses() to
determine, if an aspect uses the trait SiteLanguageAwareTrait. This
however would only work on classes that use this trait, not on classes
that inherit the class with the trait (For example XClasses).
To fix this, a SiteLanguageAwareInterface was introduced, which is
checked against.

Resolves: #87380
Releases: master
Change-Id: I55e90d40ec592dc15564c49f9437e7a2640ffb0d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61893
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>