[BUGFIX] Make EXT:backend independent of EXT:install
EXT:backend generates a URL to EXT:install, but has no dependency. This
can lead to composer setups with EXT:backend but without EXT:install,
leading to an exception while generating route.

Therefore a suggest is added as a hint to install also EXT:install.
Existing code is wrapped with an try / catch in order to work under both
circumstances.

Resolves: #89419
Releases: master, 9.5
Change-Id: Idedd299f218ae8f7ba9938800d2ac4a0c4e5614b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62261
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>