[BUGFIX] Unify return types for PDO drivers
With this change custom drivers for PDO get introduced, to unify the return
types from the database. Main topic is, to get string instead of resource
e.g. on pdo_ drivers for BLOB-fields.

Furthermore the test script for local testing gets expanded with the ability
to test on different database driver (if wished) and the option to test
lastest version of microsoft mssql server 2017.

Resolves: #88272
Releases: master, 9.5
Change-Id: I5d2cd2ccb7646f389219754be72ae6a02ae79559
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62003
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>