[BUGFIX] Use GeneralUtility for initialization of PageRenderer
The PageRenderer is set via GeneralUtility::setSingletonInstance()
by TSFE. This circumvents the DI concept and has to be removed in
the long term. It leads to the situation where makeInstance()
returns a different instance of the PageRenderer than the global DI
does.

To restore the functionality we instanciate the PageRenderer with
makeInstance() for now and will follow-up with another patch later
on to remove setSingletonInstance() usages outside of tests.

Resolves: #89445
Releases: master
Change-Id: I08846cb6592f40e5853f1ca9476969d64b7c4cce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62033
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Susanne Moog <look@susi.dev>