[TASK] Use array as service chains instead of comma-separated list
The serviceChain (services that have been requested before)
are currently added as comma-separated list (string), however,
every time when GeneralUtility::makeInstanceService() is called,
the serviceChain parameter is trimExplode'd to an array, which
leads to unnecessary CPU time when multiple services are queried.

The compat behaviour of handing in a string is now deprecated,
and in TYPO3 v11 only arrays are accepted.

Resolves: #89579
Releases: master
Change-Id: Id2424907b0d6c62fc1f249bcdd23f13d45de9893
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62043
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Oliver Hader <oliver.hader@typo3.org>