[FEATURE] Introduce palette descriptions
A new TCA property "description" is available to be added to
palette configurations. It allows to provide additional text,
which will be rendered by FormEngine between the palette label
and the corresponding fields. It is similar to the TCA field
description, introduced in #85410.

This will support editors as specific field usages can be
clarified directly in the UI.

In contrast to the palette label, the new description property
can not be overwritten on a record type basis.

Resolves: #89507
Releases: master
Change-Id: I4390e13830d359d54c37a347b91960de5639da4c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62109
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>