[TASK] Deprecate \TYPO3\CMS\Extbase\Mvc\Controller\AbstractController
The AbstractController is an internal class which never
really had any functionality than to provide some basic
method for the ActionController.

If users implemented their own controllers and
extended the AbstractController directly, the further
execution led to an exception. Unless people used the
AbstractController and implemented the missing logic
of the ActionController themselves, they didn't have
a functional controller, therefore it's safe to assume
that nobody extends the AbstractController directly.

To be on the safe side however, the AbstractController
is properly deprecated and will not be removed until
TYPO3 11.0.

People are - and always have been - encouraged to
extend the ActionController.

Releases: master
Resolves: #89554
Change-Id: I86a45621f4d32249f94deec0d932a930d3a9824f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62162
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>