[BUGFIX] Do not add entries to cache_treelist when logged into BE
When accessing cache_treelist / cObj->getTreeList()
this information should not be cached at all when logged
into the backend at the same time. This would then result
in wrong lists of IDs.

As treelist is conceptually wrong on a various number of levels,
a short workaround is added so the cache is never written
when accessing the Frontend and having a valid $GLOBALS['BE_USER']
object.

Resolves: #87540
Releases: master, 9.5
Change-Id: I1a2f0da3fffd7790dcf1988c76f6fc68899241d4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62172
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Daniel Gorges <daniel.gorges@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Daniel Gorges <daniel.gorges@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>