[BUGFIX] Respect config.intTarget and extTarget in default ParseFunc
Fluid Styled Content's "lib.parseFunc" always sets the extTarget (target for
external URLs) to the constant, but never considers config.extTarget anymore.

Same goes for the internal target, where target is never set, as
the target parameter is already filled.

The patch adds some minor inline information as well as
the proper values.

In the future, it would make sense to use the same target values
for config.*target (which is for ALL links and not just the ones in parseFunc)
which should be populated by the constants, and not have the constants
interfer into lib.parseFunc anymore.

Resolves: #88325
Releases: master, 9.5
Change-Id: Ibfa6578900f167ea465476865c67064878e31bac
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62281
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>