[BUGFIX] Streamline SoftReferenceIndex references
This patch streamlines the return values for softref type typoscript and typoscript_tag. Furthermore tests for the API method findRef() are added to ensure same return results.

Furthermore  a non-reachable part in findRef_typolink_tag is removed. As the incoming content is split by <a> tags, the <LINK> tag condition within this function will be never fulfilled.

Resolves: #88207
Related: #87652
Releases: master, 9.5
Change-Id: Ifca40c5d0e049c5b5d92a507a9d5ec0249e9953e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62196
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>