[TASK] Do not add target attribute if lightbox is enabled
If config.fileTarget is set, per example to "_blank", the target is
set as fallbak on file link. The target is also added on click enlarge
link but it is not wanted if lightbox is enabled.

This patch ensure that the target attribute is not printed when
lightbox is enabled.

Resolves: #89613
Releases: master, 9.5
Change-Id: Ib40359939a7d8689f43b3418b7b5865794046f1d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62248
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Sascha Rademacher <sascha.rademacher+typo3@gmail.com>
Tested-by: Henning Liebe <h.liebe@neusta.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Sascha Rademacher <sascha.rademacher+typo3@gmail.com>
Reviewed-by: Henning Liebe <h.liebe@neusta.de>
Reviewed-by: Benni Mack <benni@typo3.org>