[DOCS] Fix RTE preset example
The example for overwriting preset of a single field for specific type
(e.g. CType) was wrong. This is now fixed.

Also All examples were modified to create a better readable list to be
read from less specific to most specific with separated examples for
syntax and real world usage.

Resolves: #88858
Releases: master, 9.5
Change-Id: I759a606360b4e612e734cf70b7b992718d683dc9
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62251
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>