[BUGFIX] Remove leading slash from backendLogo path
As the extension configuration field `backendLogo` isn't resolved
by `TYPO3\CMS\Backend\Controller\LoginController::getUriForFileName()`,
it is in contrast to other fields (e.g. loginLogo or loginBackgroundImage)
not possible to add a path with a leading slash.

As this is neither explicit allowed nor disallowed, this unifies the
behavior by removing a possible leading slash from the defined path
before it gets resolved as it's already the case for the other fields.

Resolves: #89636
Releases: master, 9.5
Change-Id: I3f7a72f1005bc403c72b837cf66e5303962a8a42
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62293
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Reviewed-by: Benni Mack <benni@typo3.org>