[BUGFIX] Correctly pass arguments to ext:felogin "postProcContent" hook
If multiple functions are present, that call the hook
$GLOBALS['TYPO3_CONF_VARS']['EXTCONF']['felogin']['postProcContent'],
only the changes of the last executed hook would be present, as each
hook in the loop will be executed with the initial value of $content, .
not the last written $content.
To fix this behaviour, the $_params must be initialized inside the loop.

Resolves: #89679
Releases: master, 9.5
Change-Id: Ie54c45d9126acd01637c19c289c9942d0452fdae
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62941
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>