[BUGFIX] Remove jumpToUrl function in LinkBrowser
This patch backports the changes in AbstractLinkBrowserController
which were added in patch #89033.

Before, the AbstractLinkBrowserController added a href=#
and an onclick=jumpToUrl() which triggered JavaScript to switch
tabs. In this patch, the onclick=jumpToUrl() function will
be removed from a tab and complete url will be set to the tab
instead.

The onclick=jumpToUrl() function caused problems when switching
tabs while resources were loading.

Resolves: #89619
Related: #89033
Releases: 9.5
Change-Id: Ib16f86ddab8ca47e492b4ef5bb730cd73b6b5220
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62344
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Richard Haeser <richard@maxserv.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Richard Haeser <richard@maxserv.com>