[TASK] Deprecate TSFE->divSection
DivSection was used to create GMENU_LAYERS functionality,
injecting global inline JavaScript right after the rendered pages'
<body> tag.

GMENU_LAYERS was removed with TYPO3 v8, however the property
was still replaced every single time in the TYPO3 Frontend when
a USER_INT was on the page. This is not needed anymore, marking
the property as unused. That's why it is deprecated.

Resolves: #89722
Releases: master
Change-Id: Ia8e3910e20bc28831ba3bdce143812f8acb9d139
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62347
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>