[BUGFIX] Remove jumpToUrl function
This patch backports the changes in AbstractLinkBrowserController
which were added in patch #89033. There were some differences
between master and the 8.7 branch, so some adjustments had to be
made to backport the patch properly.

Before, the AbstractLinkBrowserController added a href=#
and an onclick=jumpToUrl() which triggered JavaScript to switch
tabs. In this patch, the onclick=jumpToUrl() function will
be removed from a tab and complete url will be set to the tab
instead.

The onclick=jumpToUrl() function caused problems when switching
tabs while resources were loading.

Resolves: #89619
Releases: 8.7
Change-Id: I75562d73ce91f72fc67ac79bf01f4b231dddbe4c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62356
Tested-by: Richard Haeser <richard@maxserv.com>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Richard Haeser <richard@maxserv.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>